+++ iphone/Squirrel/DataSet.h
@@ -15,7 +15,7 @@
 	NSDate *lastUpdated;
 	NSMutableArray *dataItems;
 	
-	float total;
+	double total;
 	
 	BOOL hydrated;
 	BOOL dirty;
@@ -23,7 +23,7 @@
 }
 
 @property (assign, nonatomic, readonly) NSInteger primaryKey;
-@property (assign, nonatomic, readonly) float total;
+@property (assign, nonatomic, readonly) double total;
 @property (copy, nonatomic) NSString *name;
 @property (copy, nonatomic) NSDate *lastUpdated;
 @property (retain, nonatomic) NSMutableArray *dataItems;
